+++ packages/cientos/src/core/shapes/Box.vue
@@ -42,7 +42,6 @@ defineExpose({
 <template>
   <TresMesh
     ref="boxRef"
-    v-bind="$attrs"
   >
     <TresBoxGeometry :args="args" />
     <slot>